requirements.txt
Django>=5.0
openai>=1.0.0
python-dotenv>=1.0.0
google-genai==1.65.0
langgraph==1.0.10
numpy<2
opencv-python>=4.10.0
ultralytics>=8.3.0
pyrealsense2>=2.56.0
